remove: exclusão de arquivos e referências desnecessários ao projeto
