+++ Program.cs
@@ -1,4 +1,3 @@
-using MinimalApi.Endpoints;
 using MinimalApi.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -7,8 +6,6 @@ builder.RegisterServices();
 var app = builder.Build();
 
 app.RegisterMiddlewares();
-
-//app.RegisterUserEndpoints();
 
 app.Run();
